feat: implement shopping cart and checkout management controllers with stock reservation and coupon validation

+++ web/app/Http/Controllers/CartController.php
@@ -247,7 +247,7 @@ class CartController extends Controller
             $randomHex = strtoupper(substr(md5(uniqid()), 0, 5));
             $orderRef = 'ORD' . $timestamp . $randomHex;
 
-            $expiryMinutes = BotSetting::where('key', 'checkout_expiry_minutes')->value('value') ?? 15;
+            $expiryMinutes = (int)(BotSetting::where('key', 'checkout_expiry_minutes')->value('value') ?? 15);
             $expiresAt = now()->addMinutes($expiryMinutes);
 
             // 3. Create Order
